Implemented jest integration testing
