+++ package.json
@@ -7,7 +7,8 @@
     "format": "prettier -w .",
     "start": "node ./src/server.js",
     "dev": "nodemon --exec yarn start",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest -i --colors --verbose --detectOpenHandles",
+    "test:watch": "jest -i --watchAll"
   },
   "repository": {
     "type": "git",
@@ -29,6 +30,7 @@
     "eslint": "^5.16.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^8.3.0",
+    "jest": "^27.0.6",
     "nodemon": "^2.0.12",
     "prettier": "^2.3.2"
   }
