+++ src/resolvers/resolvers.js
@@ -1,0 +1,7 @@
+const resolvers = {
+  Query: {
+    sayHi: () => 'Hello',
+  },
+}
+
+module.exports = resolvers
