+++ src/server.js
@@ -1,17 +1,7 @@
 const { ApolloServer } = require('apollo-server')
-const gql = require('graphql-tag')
 
-const typeDefs = gql`
-  type Query {
-    sayHi: String!
-  }
-`
-
-const resolvers = {
-  Query: {
-    sayHi: () => 'Hello',
-  },
-}
+const typeDefs = require('./typeDefs/typeDefs')
+const resolvers = require('./resolvers/resolvers')
 
 const server = new ApolloServer({
   typeDefs,
