+++ src/typeDefs/typeDefs.js
@@ -1,0 +1,9 @@
+const gql = require('graphql-tag')
+
+const typeDefs = gql`
+  type Query {
+    sayHi: String!
+  }
+`
+
+module.exports = typeDefs
